backend/requirements.txt
fastapi[all]
uvicorn[standard]
motor
pandas
groq
python-dotenv